setting up for content_types and custom columns
