+++ src/Console/Commands/IgorWatchCommand.php
@@ -54,8 +54,10 @@ class IgorWatchCommand extends Command
         $types = $this->files->directories($staticPath);
         foreach ($types as $type_path) {
             $type = basename($type_path);
-            $posts = $this->files->directories("$staticPath/$type");
-            foreach ($posts as $post) {
+            //$this->igor->createContentType($type);
+            //$this->igor->createCustomColumnTypes($type);
+            $contents = $this->files->directories("$staticPath/$type");
+            foreach ($contents as $post) {
                 $igor = new Igor($post, new IgorRepository);
                 $igor->reAnimate();
                 $igor->reAnimateAssets();
