+++ src/Models/Asset.php
@@ -11,9 +11,9 @@ class Asset extends Model
     /**
      * Get all of the posts that are assigned this asset.
      */
-    public function posts()
+    public function content()
     {
-        return $this->morphedByMany('App\Post', 'assetable');
+        return $this->morphedByMany('App\Content', 'assetable');
     }
 
     public function type()
