+++ src/Models/Category.php
@@ -11,8 +11,8 @@ class Category extends Model
     /**
      * Get all of the posts that are assigned this tag.
      */
-    public function posts()
+    public function content()
     {
-        return $this->morphedByMany('App\Post', 'categorable');
+        return $this->morphedByMany('App\Content', 'categorable');
     }
 }
