+++ src/Models/ColumnType.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Jeremytubbs\Igor\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class ContentType extends Model
+{
+    protected $guarded = ['id'];
+
+    public function contents()
+    {
+        return $this->hasMany('Jeremytubbs\Igor\Models\Content');
+    }
+}
