+++ src/Models/Content.php
@@ -8,8 +8,37 @@ class Content extends Model
 {
     protected $guarded = ['id'];
 
-    public function publishable()
+    protected $dates = ['published_at'];
+
+    /**
+     * Get all of the tags for the post.
+     */
+    public function tags()
     {
-        return $this->morphTo();
+        return $this->morphToMany('Jeremytubbs\Igor\Models\Tag', 'taggable')->withTimestamps();
+    }
+
+    /**
+     * Get all of the posts categories.
+     */
+    public function categories()
+    {
+        return $this->morphToMany('Jeremytubbs\Igor\Models\Category', 'categorable')->withTimestamps();
+    }
+
+    /**
+     * Get all of the posts assets.
+     */
+    public function assets()
+    {
+        return $this->morphToMany('Jeremytubbs\Igor\Models\Asset', 'assetable')->withTimestamps();
+    }
+
+    /**
+     * Get the user.
+     */
+    public function user()
+    {
+        return $this->belongsTo('App\User');
     }
 }
