+++ src/stubs/Models/Content.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App;
+
+use Jeremytubbs\Igor\Models\Content as IgorContent;
+
+class Content extends IgorContent
+{
+    use \Jeremytubbs\Igor\Traits\SluggerTrait;
+
+}
