--- src/stubs/Models/Page.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App;
-
-use Jeremytubbs\Igor\Models\Page as IgorPage;
-
-class Page extends IgorPage
-{
-    use \Jeremytubbs\Igor\Traits\SluggerTrait;
-
-}
