--- src/stubs/postModel.stub
@@ -1,11 +1,0 @@
-<?php
-
-namespace {{namespace}};
-
-use Jeremytubbs\Igor\Models\Post as Igor{{class}};
-
-class {{class}} extends Igor{{class}}
-{
-    use \Jeremytubbs\Igor\Traits\SluggerTrait;
-
-}
